Moved Texture and Shader load to TextureLoader and ShaderLoader

+++ OrbeusEngine/Source/Core/Input.cpp
@@ -10,7 +10,7 @@ static Input::MouseState lastMouseState;
 
 void scroll_callback(GLFWwindow* window, double xoffset, double yoffset)
 {
-	currentMouseState.scroll = yoffset;
+	currentMouseState.scroll = (float)yoffset;
 }
 
 void Input::setWindow(GLFWwindow* window)
